# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "示例API文档.",
        "title": "示例API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8089",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/assignRole": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "给用户分配角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "给用户分配角色",
                "parameters": [
                    {
                        "description": "用户ID与角色编号",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AssignRoleModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result-any"
                        }
                    }
                }
            }
        },
        "/user/createByEmail": {
            "post": {
                "description": "邮箱注册用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "邮箱注册用户",
                "parameters": [
                    {
                        "description": "邮箱与密码",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserEmailRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result-model_UserModel"
                        }
                    }
                }
            }
        },
        "/user/delete/:id": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "刪除用戶",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "刪除用戶",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result-any"
                        }
                    }
                }
            }
        },
        "/user/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户管理列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户管理列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "排序类型：asc、desc",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页码",
                        "name": "pageNo",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "姓名",
                        "name": "realName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态: 1-启用，2-禁用",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result-model_PageData-model_UserModel"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "使用邮箱登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "使用邮箱登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserLoginModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.AssignRoleModel": {
            "type": "object",
            "required": [
                "roleCode",
                "userId"
            ],
            "properties": {
                "roleCode": {
                    "description": "角色编码",
                    "type": "string"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "integer"
                }
            }
        },
...
        "model.PageData-model_UserModel": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "总记录数",
                    "type": "integer"
                },
                "data": {
                    "description": "分页数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserModel"
                    }
                },
                "pageNo": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页数量",
                    "type": "integer"
                }
            }
        },
...
        "model.UserEmailRegister": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
        "model.UserLoginModel": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
        "model.UserModel": {
            "type": "object",
            "properties": {
                "createAt": {
                    "description": "创建时间",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.LocalDateTime"
                        }
                    ]
                },
                "createBy": {
                    "description": "创建人",
                    "type": "integer"
                },
                "deleted": {
                    "description": "是否删除 1-否，2-是",
                    "type": "integer"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "主键id",
                    "type": "integer"
                },
                "realName": {
                    "description": "姓名",
                    "type": "string"
                },
                "roleCode": {
                    "description": "角色编号",
                    "type": "string"
                },
                "status": {
                    "description": "状态: 1-启用，2-禁用",
                    "type": "integer"
                },
                "updateAt": {
                    "description": "更新时间",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.LocalDateTime"
                        }
                    ]
                },
                "updateBy": {
                    "description": "更新人",
                    "type": "integer"
                },
                "userName": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "result.Result-any": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码，200表示成功，其它为失败",
                    "type": "string"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                },
                "timeStamp": {
                    "description": "时间戳",
                    "type": "integer"
                }
            }
        },
...
        "result.Result-model_PageData-model_UserModel": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码，200表示成功，其它为失败",
                    "type": "string"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.PageData-model_UserModel"
                        }
                    ]
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                },
                "timeStamp": {
                    "description": "时间戳",
                    "type": "integer"
                }
            }
        },
        "result.Result-model_UserModel": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码，200表示成功，其它为失败",
                    "type": "string"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.UserModel"
                        }
                    ]
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                },
                "timeStamp": {
                    "description": "时间戳",
                    "type": "integer"
                }
            }
        },
        "types.LocalDateTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}